This is my second attempt at addressing the speed issue in the comments on #1915.

In this PR:
* Do cache operations using batch methods
* Keep the _verbose JSON generation, but prefetch the stages in one big batch
* Default update_cache to True

Since we are still loading stages, the performance improvements are not as good as my first approach:
* Original code: cache miss 7s, cache hit 2s
* This code: cache miss 5.5 to 7s, cache hit 1.6 to 1.9s

I think the speedup in this PR is worth merging, but we need to follow up with another change that uses the _basic() conversion function again, even if that requires some client-side changes and additional requests on the myfeatures page.